+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase length penalty for primary key detection
 - Update to Scala 2.13
 - Move property set configuration to params object
+- Only generate tuple schemas on multiple observations
 
 ### Fixed
 - Allow schema replacement to replace the entire schema